FIX116 - Rename the common folder to shared
