+++ src/shared/decorators/autobind.ts
@@ -1,12 +1,10 @@
 export default (_target: unknown, _name: string, descriptor: PropertyDescriptor): PropertyDescriptor => {
   const original = descriptor.value;
-  const adjusted: PropertyDescriptor = {
+  return {
     configurable: true,
     enumerable: false,
     get() {
       return original.bind(this);
     },
   };
-
-  return adjusted;
 };
