+++ src/shared/enums/http.ts
