+++ src/shared/errors/bad-request.ts
@@ -1,0 +1,8 @@
+import BaseError from '@src/shared/errors/base';
+import { HttpStatusCodes } from '@src/shared/enums';
+
+export default class BadRequestError extends BaseError {
+  public constructor(public message: string) {
+    super('BadRequestError', HttpStatusCodes.BAD_REQUEST, message);
+  }
+}
