+++ src/shared/errors/base.ts
