+++ src/shared/errors/validation.ts
@@ -1,5 +1,5 @@
-import BaseError from './base';
-import { HttpStatusCodes } from '../enums/http';
+import BaseError from '@src/shared/errors/base';
+import { HttpStatusCodes } from '@src/shared/enums';
 
 export default class ValidationError<DataType> extends BaseError {
   public constructor(public message: string, public data: DataType) {
