+++ src/shared/interfaces/database.ts
